components seperated
